requirements.txt
aiohttp==3.6.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
discord.py==1.3.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
numpy==1.18.1
praw==6.5.1
prawcore==1.0.1
pylint==2.4.4
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
update-checker==0.16
urllib3==1.25.9
websocket-client==0.57.0
websockets==8.1
wrapt==1.11.2
yarl==1.4.2